add a test and dry-run to qmk generate-api

+++ lib/python/qmk/cli/generate/api.py
@@ -11,6 +11,7 @@ from qmk.info import info_json
 from qmk.keyboard import list_keyboards
 
 
+@cli.argument('-n', '--dry-run', arg_only=True, action='store_true', help="Don't write the data to disk.")
 @cli.subcommand('Creates a new keymap for the keyboard of your choosing', hidden=False if cli.config.user.developer else True)
 def generate_api(cli):
     """Generates the QMK API data.
@@ -36,10 +37,14 @@ def generate_api(cli):
         keyboard_readme_src = Path('keyboards') / keyboard_name / 'readme.md'
 
         keyboard_dir.mkdir(parents=True, exist_ok=True)
-        keyboard_info.write_text(json.dumps({'last_updated': current_datetime(), 'keyboards': {keyboard_name: kb_all['keyboards'][keyboard_name]}}))
+        keyboard_json = json.dumps({'last_updated': current_datetime(), 'keyboards': {keyboard_name: kb_all['keyboards'][keyboard_name]}})
+        if not cli.args.dry_run:
+            keyboard_info.write_text(keyboard_json)
+            cli.log.debug('Wrote file %s', keyboard_info)
 
-        if keyboard_readme_src.exists():
-            copyfile(keyboard_readme_src, keyboard_readme)
+            if keyboard_readme_src.exists():
+                copyfile(keyboard_readme_src, keyboard_readme)
+                cli.log.debug('Copied %s -> %s', keyboard_readme_src, keyboard_readme)
 
         if 'usb' in kb_all['keyboards'][keyboard_name]:
             usb = kb_all['keyboards'][keyboard_name]['usb']
@@ -53,6 +58,16 @@ def generate_api(cli):
             usb_list['devices'][usb['vid']][usb['pid']][keyboard_name] = usb
 
     # Write the global JSON files
-    keyboard_list.write_text(json.dumps({'last_updated': current_datetime(), 'keyboards': sorted(kb_all['keyboards'])}))
-    keyboard_all.write_text(json.dumps(kb_all))
-    usb_file.write_text(json.dumps(usb_list))
+    # Note: This is 70mb of JSON as of Jan 2021, and will only get bigger. If the size becomes a problem we can switch to a more memory efficient process that doesn't let us dry run as easily.
+    keyboard_list_json = json.dumps({'last_updated': current_datetime(), 'keyboards': sorted(kb_all['keyboards'])})
+    keyboard_all_json = json.dumps({'last_updated': current_datetime(), 'keyboards': sorted(kb_all['keyboards'])})
+    usb_json = json.dumps(usb_list)
+
+    if not cli.args.dry_run:
+        keyboard_list.write_text(keyboard_list_json)
+        cli.log.debug('Wrote file %s', keyboard_list)
+        keyboard_all.write_text(keyboard_all_json)
+        cli.log.debug('Wrote file %s', keyboard_all)
+        usb_file.write_text(usb_json)
+        cli.log.debug('Wrote file %s', usb_file)
+        cli.log.info('Wrote API data to %s', api_data_dir.resolve())
